Move backwardcompat package to compat module
