+++ pip/commands/search.py
@@ -4,7 +4,7 @@ import textwrap
 from pip.basecommand import Command, SUCCESS
 from pip.util import get_terminal_size
 from pip.log import logger
-from pip.backwardcompat import xmlrpclib, reduce, cmp
+from pip.compat import xmlrpclib, reduce, cmp
 from pip.exceptions import CommandError
 from pip.status_codes import NO_MATCHES_FOUND
 from pip._vendor import pkg_resources
