+++ pip/compat.py
@@ -119,7 +119,7 @@ def get_path_uid(path):
     Does not follow symlinks:
         https://github.com/pypa/pip/pull/935#discussion_r5307003
 
-    Placed this function in backwardcompat due to differences on AIX and
+    Placed this function in compat due to differences on AIX and
     Jython, that should eventually go away.
 
     :raises OSError: When path is a symlink or can't be read.
