+++ pip/vcs/git.py
@@ -5,7 +5,7 @@ from pip.util import call_subprocess
 from pip.util import display_path, rmtree
 from pip.vcs import vcs, VersionControl
 from pip.log import logger
-from pip.backwardcompat import url2pathname, urlparse
+from pip.compat import url2pathname, urlparse
 
 urlsplit = urlparse.urlsplit
 urlunsplit = urlparse.urlunsplit
