+++ pip/vcs/subversion.py
@@ -1,6 +1,6 @@
 import os
 import re
-from pip.backwardcompat import urlparse
+from pip.compat import urlparse
 from pip.util import rmtree, display_path, call_subprocess
 from pip.log import logger
 from pip.vcs import vcs, VersionControl
