+++ tests/functional/test_install_user.py
@@ -7,7 +7,7 @@ import textwrap
 
 from os.path import curdir, isdir, isfile
 
-from pip.backwardcompat import uses_pycache
+from pip.compat import uses_pycache
 
 from tests.lib.local_repos import local_checkout
 from tests.lib import pyversion
